Switch prerelease info for build info when handing off our version to Scout.
Why? We want to use our version as a Docker tag, and DockerHub doesn't allow "+" in a tag -- so we use the semver prerelease instead of the build. But Scout wants to see semver build info as a marker for "don't include this in analytics." Sooooo we swap here.

(A Better Answer™ might be to modify the Docker tag instead...)