Added daemonset feature to helm
This will fix issue #309

Fixed typo

Revert changes to Chart.yaml file

Bump Helm chart version to 0.40.0.

adding correct protocol for upstream service

For Ambassador to use mTLS correctly, the upstream should specify the right protocol to use, in this case, https. Setting the "tls" option is not enough. Tested today with 0.40

adding correct protocol for upstream service

For Ambassador to use mTLS correctly, the upstream should specify the right protocol to use, in this case, https. Setting the "tls" option is not enough. Tested today with 0.40

Bump Helm chart version to 0.40.0.