Run: make build
This updates a couple of versions in `go.mod`.  It also generates
`go.sum`; in principle I should also be commiting that.  But I'm not
doing that yet, because for

    github.com/datawire/teleproxy v0.0.0-20181207190820-b9379890f5d0

my (Parabola GNU/Linux-libre) laptop calculates the sum as

    h1:pt7nPoBZzSNv8ZLPfOxcCsm2RptG18bc08fCHZjLQcY=

but CI Ubuntu, CI macOS, and MacInCloud calculate it as

    h1:svxbWqiNeTzdbn0y9pNOZzrILKK//YMpwKwKIerDot4=

I've tried clearing all caches on my laptop and re-downloading it, but
I still get that checksum.

So... maybe I'm getting MITM'd, but this would be a funny target for
that.  I suspect that maybe it's down to a difference in `git`
behavior?  IDK.  But I'm not committing `go.sum` until I figure it
out.