Makefile: Fix `make deploy` on macOS
We set up a `kubectl port-forward` so that localhost:31000 forwards to the
Docker registry running in the cluster, and do a `docker push` to
"localhost:31000/whatever".  That's all fine and dandy on operating
systems where dockerd runs.

But dockerd won't run on macOS, so Docker for Mac runs dockerd in a
GNU/Linux VM.  Which means that dockerd's "localhost" a VM, not the macOS
host that `kubectl port-forward` set up forwarding for.

So, to work around this:

 1. Docker for Mac exposes the macOS host as `host.docker.internal`.  So,
    for the local `docker` commands, on macOS, we need to set the registry
    to "host.docker.internal:31000" instead of "localhost:31000".

 2. But then dockerd will try to talk HTTPS to it, instead of HTTP; it
    speaks HTTPS to "secure registries", and HTTP to "insecure
    registries", where the list of insecure registries is configurable,
    but is hard-coded to include "localhost".  So, we have to tell macOS
    developers to configure dockerd to consider
    "host.docker.internal:31000" to be insecure.

 3. But then that would set

        image: host.docker.internal:31000/ambassador-pro:whatever

    in the YAML.  So use a $(foreach) to set up a local scope in the
    Makefile where we set it back to "localhost".