Add a cached container image
This commit splits the current Dockerfile to a cached Dockerfile
which installs all the required packages and dependencies for
Ambassador, and another one which imports from the cached
container image and copies in current Ambassador code and
configures that as well.

This has been done to speed up the dev loop and to avoid
unreliable Alpine mirrors.