amb-sidecar: Prepare to split OAuth2 out of the main HTTP handler [ci-skip]
This prepares for splitting handler_oauth2.go in to 2 separate files,
by doing a "copy" and a "rename".  By doing these as a separate step
from actually modifying either file, it both makes rebasing easier,
and makes it easier to make sense of the diff, for code-review.

If you see this commit in `git bisect`, then `skip` it; the
`cmd/amb-sidecar/oauth/app/oauth2handler` package is known to be bogus
at this point.