OAuth2 filter: Allow 'scope' claims to be actual lists
We expect the 'scope' claim in the returned JWT to be a string with a
whitespace-separates list of scopes; it seems UAA is returning the list of
scopes as an actual array.  My reading of the specs is that UAA is
operating out-of-spec, but that doesn't mean we shouldn't support it.

Also, handle weird types better than panicing.