max_load.go: Don't leak so many file descriptors
The Attacker keeps open 1000 (by default; the number is configurable) idle
connections, kept by its internal http.Client.  We need to call
CloseIdleConnections() on the client when we're done with the Attacker,
but the vegeta.Attacker API keeps the http.Client hidden from us and gives
us no way to do that.

When the connection objects get garbage collected, the file descriptors
get closed, but because they don't take up much memory, them getting
collected reliably/routinely enough is a crap-shoot.  If you finish with 1
Attacker and then create a new one, you'll probably have 2000 open
connections.  Without GC, every new Attacker adds another 1000 open
connections.  With GC, the number will be less than that, but not by a ton.

This leads us to hit resource limits on the number of open file
descriptors.

But, there's no reason we need to make a new Attacker every time.  We can
just have a global

    var attacker = vegeta.NewAttacker()

that we reuse.