Allow DNS resolvable names for endpoint routing
Before this commit, if an Ambassador Mapping has a service set as
`qotm.default.svc.cluster.local`, then it is not resolved to
endpoints. This is an issue in cases where long resolvable
hostnames are used which span namespaces instead of short namespace
specific names.

This commit tries to resolve such a name and then does a reverse
lookup in Kubernetes services for the resolved IP to allow for
endpoint routing.