Merge tests/cluster/oidc/ in to tests/cluster/oauth-e2e/, clean up
I spent too long trying to clean up the Go tests/cluster/oidc tests,
to be less tied to implementation details.  This mostly consisted of
implementing browser behavior in Go.  Then I realized that was dumb, and
doing this with an actual browser (and browser scripting) is the way to
go, even if JS is miserable.