Remove x_forwarded_proto_redirect infinite loop
This commit removes a bug where when you set the ambassador module
config x_forwarded_proto_redirect to true it would generate a
configuration that caused an infinite redirect loop.

The main cause of this issue was that the if you set the
x_forwarded_proto_redirect to true it would only create an envoy
listner with the default route rather than still appending the
desired custom routes specified in the Kubernetes services annotations.

The default route now doesn't contain a redirect that does a
[https_redirect](https://www.envoyproxy.io/docs/envoy/latest/api-v2/api/v2/route/route.proto#route-redirectaction)
due to it causing an infinite loop of redirects as it
would substitue the scheme portion of the URL for https and redirect to
that URL even if that scheme was already https.

Removed the check for the `listener.redirect_listener` before building
the routes so the routes would always be built and defult route would be
overwriten if routes have been defined wether or not
`listener.redirect_listener` had been set or not. Added a condition to
set the `self.require_tls = None` only if `listener.redirect_listener`
had not be set.

This change now means that Ambassador will add the `"require_tls":
"EXTERNAL_ONLY"` configuration to the envoy virtal host if the
Ambassador module `x_forwarded_proto_redirect` is set to true forcing
all external requests to the gateway to use https.

Corrects issue: https://github.com/datawire/ambassador/issues/1571