build-aux-push: Do a fetch first
If a `git-subtree-split:` line refers to a commit that we don't have
locally (because it has been garbage collected, because
build-aux.git isn't a remote in .git/config), then `git subtree`
will do the wrong thing (instead of just aborting, because crappy
error handling), and potentially push proprietary code to
build-aux.git.  So do a fetch first, to guarantee that build-aux
commits exist locally.