base-envoy: Always store a non-stripped binary
Even if we're going to strip it for the final Ambassador release.
It's cheap-enough to just strip locally before building the final
Ambassador image (since it doesn't involve the network).

This way, we always have the non-stripped binary somewhere.  This is
important because if Envoy crashes, to make sense of the stack trace, you
need to run it through `tools/stack_decode.py`, and stack_decode.py
requires a non-stripped binary for reference.  We should publish this
somewhere, so that we have the actual binary, instead of me re-compiling
Envoy and hoping the symbol addresses are the same.