build-aux: Upgrade teleproxy and kubeapply
The ancient teleproxy was giving me problems on my home network, so
upgrade it.

I went with the current 'master' instead of the most recent tag, so that
it would include abaf7a3de6f3ef33e841166760c57138df0c8d5a (libk8s), and I
wouldn't have to deal with weird versioning stuff to get it to build.

And then, for teleproxy (but not kubeapply), I bumped that to
4ad7074eeea12dd20ec3c1e8927bd5ca5b004fb6 (which is not yet on master), so
that it would work with multi-port services (like Ambassador).