go-mod.mk: Drop the public go.bin.rule function
With the Go 1.13 improvements, it stopped being useful to outside callers.
The "right" thing to do would be to make the Go 1.13 version private, and
always export the Go 1.11-1.12 version (which works with Go 1.13).  But
the only user of it is APro's Lyft RL binaries, and no one uses those, so
I'm just making the whole thing private.