rm -rf cmd/loadtest-locust-slave
Go seems to be getting confused by github.com/ugorji/go/codec, which is
pulled in by the Locust slave.  The Locust slave is essentially dead code,
I was OK keeping it around, since it may be useful, and didn't really
cause any drag.  Now it causes drag.  Remove it.

    $ go mod why github.com/ugorji/go/codec
    # github.com/ugorji/go/codec
    github.com/datawire/apro/cmd/loadtest-locust-slave
    github.com/myzhan/boomer
    github.com/ugorji/go/codec