Revert capabilities on envoy binary
Some environments don't like capabilities enabled binaries. This rolls
back #1720 which adds the net_bind_service capability directly to Envoy.

I will work on a different version that detects if the capability would
cause a problem, and to use a wrapper if so.

I have left the parts of the test suite in that let us specify the
capabilities block in the pod template since I will use them in the next
version of this.