port over tls module fields to tlscontext
this commit adds in all fields documented in:
`docs/reference/core/tls.md` into TlsContext, it turns out this was
only missing redirect_cleartext_from. however I want to make it clear
this should be all of them at this point. So TlsContext should be the
thing that is used from this point forward.