Increase default regex_max_size to 200
The regex that we generate for the simplest knative service has
an RE2 maximum size of more than 100, which is Envoy's default.
This causes Ambassador to either crash or ignore the new config.

This limit is being hit by non-knative users as well.

This commit increases the default regex_max_size to 200 which
should be enough for our generated regex.