Add a capabilities wrapper
This wrapper is given the CAP_NET_BIND_SERVICE capability so that it can
run Envoy with that bit set. This will let Envoy run on a low port as a
non privileged user.

Envoy needs to be able to inherit the capability, but I've verified that
this does not suffer from the same problem of not being able to run when
that capability is explicitly dropped.

In a separate PR I will update the entrypoint so that it calls the
wrapper if the capability is available, and Envoy otherwise