webui: Add a tracker Ambassador logo
Bjorn:
> @LukeShu When you build the better UI, please add an Ambassador
> logo, partly to make it look good, and partly as a tracker. The logo
> should be loaded from our website and it should be loaded through a
> url that includes the Ambassador id. In other words, something like
> https://getambassador.io/static/ambassador-logo.png?12345654321
> where the 12345654321 is the Ambassador id. That will put the id in
> our website access logs and then we can use that to determine who is
> installing Ambassador, etc. Thanks.

"When you build the better UI, please add an Ambassador logo, partly to make it look good, and partly as a tracker. The logo should be loaded from our website and it should be loaded through a url that includes the Ambassador id. In other words, something like https://getambassador.io/static/ambassador-logo.png?12345654321 where the 12345654321 is the Ambassador id. That will put the id in our website access logs and then we can use that to determine who is installing Ambassador, etc. Thanks.