fix use of requestUpdate in dashboard component
this fixes the user of requestUpdate in dashboard component
(getting us to not use it), since it can cause problems for other
browsers, and there's no reason for us to maintain this code.

this also prevents the need to null check in snapshot.js + dashboard.js
since we just populate a snapshot wrapper always.