start documenting some common gotchas in the ui
the goal of this PR is to introduce some common gotchas inside a
document that can be used as a source doing review, or as somone
is writing a component.

NOTE: this doesn't actually fix any instances of this. This is just
the gotchas I found inside of `tabs.js`. there will be following prs
for fixing these in tabs.js, and others.