stop depending on array ordering in tabs.js
  This was documented in: https://github.com/datawire/apro/pull/552

the array order can change since the dom is mutable, and techincally
many other scripts can change it. as such we employ a fix of #1/#3.

First we don't try to depend on ordering. assuming if someone changes
the dom order, they want it changed. we shouldn't make an assumption
of ordering in tabs.js

Secondly we use the tab name to ensure we always render the correct
tab as active, even if the dom order changes.