acmeclient: "Gracefully" handle it when we don't have Redis
Presently, what happens if we get to requesting a certificate but we don't
have Redis is that we crash with SIGSEGV.

In the launcher, we could avoid launching the ACME client if we don't have
Redis, but then we wouldn't get defaults fill or status fill for non-ACME
Hosts.

We could just bail after rectifyPhase1 if we don't have Redis, but then
the user would have no indication why ACME wasn't working.

We could instead record an error for ACME Hosts after rectifyPhase1, but
then we might interfere with valid Hosts that would have still been good
for a while, during a botched upgrade or something.

So instead, we go through as much of the process as we can, until we
actually need Redis, and then record an error on the Hosts that would need
Redis.