acmeclient: Have the leader election try to re-aquire leadership if we lose
If we are the leader, then someone else becomes the leader WITHOUT OUR POD
DYING (perhaps because we were CPU-starved, as can be simulated with
SIGSTP), then we never try to become the leader again.

I'd been under the impression that leaderElector.Run() had done that
looping, but I was mistaken in that impression.

Thanks to @ark3 for catching this!