Move getambassador.io/config annotation unfolding to happen in watt
This is non-prioritized work, happening not because I'm spending
Datawire time on it, but because I'm spending personal time on it
because I'm not able to sleep because "move annotation unfolding to
happen in watt" is an evil clown chasing me through my dreams.  (And
thus this commit isn't coming from Datawire-employee
lukeshu@datawire.io, it's coming from OSS-contributor
lukeshu@lukeshu.com)

I can't explain how an abstract idea such as "move annotation
unfolding to happen in watt" manifests as an evil clown in a dream.
There isn't a visual to go with that.  It's an abstract non-corporeal
clown, that you just know is there, hunting you.  I do not, and have
not ever, had a fear of clowns, which makes the whole thing weirder.
However, I can say that Ambassador custom resources manifest as short
(5', tops), surly men who bump in to you in convenience stores with
narrow aisles.

I'm not entirely sure whether this is mergable in its current state.
We lose some of diagd's tracking of whether a resource came from an
annotation, and thus screw up the "name.namespace.N" identifiers.
Anyway, having [robbed the clown of his joy][1], I can sleep.

[1]: https://www.youtube.com/embed/Ngk3VX405GE?start=137&end=162