webui: If DEV_WEBUI_SNAPSHOT_HOST is set, also proxy diagd-needing requests
There are 2 requests that might need to talk to diagd:

 1. GET /edge_stack/api/snapshot
 2. POST /edge_stack/api/log-level

However, it was only proxying the first one to the amb-sidecar that
actually has a diagd to talk to.  This was causing problems for people
trying to use ui-dev-loop option 3.  Proxy them both.

And also, use httputil.ReverseProxy to do that; it means I have a lot less
code to copy/paste around.  I don't think Rafi or I knew about
httputil.ReverseProxy when he wrote the original code, or when I reviewed
it.