Add note about the status of the Ingress api
networking.k8s.io is the more recent api for the `Ingress` resource. I
bet Kubernetes will eventually remove the support for
`ingress.extensions` like they did with `Deployment` resource so we
should make sure to note why we are using `extensions` in the docs.

The RBAC was also wrong for supporting `networking.k8s.io` so that has
been fixed too

Signed-off-by: Noah Krause <krausenoah@gmail.com>