python/tests/t_*: Use the "super().manifests()"-at-the end idiom
In today's Ambassador-U, Flynn explained that the proper idiom is to
include new manifests at the beginning; eg.

        return """
    MY-YAML
    """ + super().manifests()

and not

        return super().manifests() + """
    MY-YAML
    """

pretty much because "namespaces".

But a bunch of the existing tests do it the discouraged way.  This
increases the odds that someone won't just figure-it-out looking at
whats's there.  So change it.