Due to some mistake, we weren't shipping the Envoy commit we said
So somehow, we ended up that the Docker image for
Envoy commit d17d947caef13f1bdd235c3fccff77814883bb46 (the image we use)
actually had an Envoy binary built from commit
3bc432257e20e07a49d6213c8f255bbeb4edd561.

Yikes!

It turns out that the difference is relatively minor (thankfully, both
contain the 1.12.2 security patches!):
 - 3bc4322 _always_ clears the route cache when getting a header-modifying
   response from ext_authz, instead of obeying clear_route_cache.  That
   makes sense, as clearing the cache works, and we haven't (yet) modified
   v2listener.py to set clear_route_cache.
 - 3bc4322 uses the old pb.lyft.ratelimit gRPC service name.  That makes
   sense, as KAT still works, but kat-server was still using the old name.

Both of those things are things that were historically true in
Ambassador's Envoy, but we had thought we changed in Ambassador 0.86.0.
So it turns out that those things didn't change in 0.86.0.  And I'm OK
with that.

But let's adjust ENVOY_COMMIT so that people (especially: future me) don't
get confused about which Envoy we're shipping.

Oh, and also update the CHANGELOG to reflect the truth.