oauth2: Improve comments
In the review for 3181b49f85a4b8eb9c23f2cf2dc491052032d8f6 I said that
after sleeping on it that I hated all of the comments I'd written the
night before, and we agreed to come back and improve the comments later.

So this is improving those comments.