Move to checking the Chart.yaml for a version bump
I was poking at the chart today and thinking about how awful the UX of
needing to push a tag just to get the chart to build and push a new
version is.

I decided that the simplest way to decide when to build and push is to
check to see if the version of the chart was updated.

The simplest way I could think to accomplish this is by taking a diff of
the Chart.yaml file and checking to see if the `version` was edited.

Travis has an environment variable `TRAVIS_COMMIT_RANGE` for storing the
range of the pull request that triggered this build. I believe using
this variable to `git diff Chart.yaml` accomplishes what I am looking
for.

Signed-off-by: Noah Krause <krausenoah@gmail.com>