filters: Consistently use X-Forwarded-Proto instead of GetScheme
I was looking at filters/handler/handler.go's requestURL() function and
noticed

 1. the complexity of it, compared to the simplicity of the version in
    oauth2handler/client/authorization_code_client/client.go.  What
    edge-cases aren't being handled by the oauth2 handler version?
 2. the call to .GetScheme(), which is known to be broken.

So, I moved it to the filterutil library, so that I could fix everything
in one place.  Then I noticed that a few lines up in filterutil was
another call to the broken .GetScheme().  So fix that too.