Prune down cmd/aes-plugin-runner{,-native}
Because of cmd/aes-plugin-runner{,-native} being copies of eachother,
there was a lot of stuff in aes-plugin-runner-native that was only needed
in the plain aes-plugin-runner, and vice-versa.  So trim those things
down.

"Ignore space change" might be useful for viewing this patch, as a couple
of "if{}" blocks get changed to always run.