Host crd: Fix insecure Route action
Fixes: https://github.com/datawire/ambassador/issues/2216

For each listener, there will be one envoy filter per server_names
and also an envoy filter with empty filter_chain.

For http, the envoy filter_chain by server_names has no effect.
This is only for SNI and https.
So for the http listener, only the first envoy filter is used.

So, when the first vhost is redirecting, all the http traffic will be redirected.
When the first vhost is routing http traffic, all the http traffic will be routed, no redirection.

The fix is checking per route, within the vhost loop,
if there is a corresponding "Host" crd for the route having the hostname.
If so, it applies the action based on the Host CRD and not based on the vhost currently processed.