kale: Clarify the phase-calculation code
This started off as me adding comments based on a conversation with Rafi.
But then I realized I think the whole thing gets clearer if I change it to
be a switch instead of an if/else ladder.

As a result of Rafi pointing out that the `len(builders) != 1` check was
really handling 2 distinct cases (`=0` and `>=2`), I now do handle the two
cases separately.  And in the >=2 case, set phase=Building, instead of
phase=Received.

Maybe check "ignore space change" when viewing this diff, one of the
branches lost a level of indentation, but didn't change.