diagd: Refactor EnvoyStats
EnvoyStats has been split into two objects:

EnvoyStats: this object has the envoy statistics at one point in time
and it is immutable. Once a client gets an instance, it can reuse it and
state is guaranteed to not change.
The instance should not be hold or cached for future use except for the
time of a request or processing some state from this state.

EnvoyStatsManager: this object is responsible to update the envoy
statistics and expose it to the clients in an immutable way.
When a client requests the statistics at a point in time, the manager
will do a deep copy of the state and return this so the state returned
to the client will not change overtime.
This object is responsible for synchronizing the state changes
internally.