Also check for `allowPrivilegeEscalation`
The Ambassador container needs both the NET_BIND_SERVICE capability and
also securityContext.allowPrivilegeEscalation to be enabled, so updating
the init script to check for both. It checks to make sure it's not
disabled (love these double negatives) so that if we're on a non docker
runtime it won't get tripped up.