Fix #97: Cannot specify podSecurityPolicies
If one want to use a custom `values.yaml` for their Helm chart,
the bundled `values.yaml` of the existing Helm chart will get
merged with the custom values.yaml.

Since the bundled `values.yaml` defines an array `[]` for
podsecuritypolicies, but the podsecuritypolicies.yaml template
expects a map for `annotation:` and `spec:`, it is not possible
to generate a valid Helm chart.

The only workaround would be to ignore/delete the bundled values.yaml
altogether.

This PR fixes this issue by using a map: `{}` as a default value for
the bundled `values.yaml`

Fixes:
- https://github.com/datawire/ambassador-chart/issues/97