(from AES) git checkout 0000074959c8f601d27489397b2cbcf0cbb3edff ambassador/api/getambassador.io/v2/Host.proto && make generate
On one branch it was removed in favor of kubebuilder.
On another branch it was modified.
When the two branches were merged, the file was deleted and the
modifications were re-done on the kubebuilder version.
When we `git revert`ed the removal, we ended up with the one from the
wrong branch; the one without the modifications.