(from AES) Set up dlog from pkg/busy
So now all commands that are part of busyambassador will have consistent
logging, as long as they use dlog.  (AND EVERYTHING SHOULD USE DLOG!)

And while we're at it, this is a good central place to make sure that it
logs the PID and which command it's running, since confusion around that
has been a source of trouble when reading the logs.