Drop support for Knative ClusterIngresses
This change removes support for Knative ClusterIngresses, which have
been superseded by Ingresses and removed from Knative entirely since
version 0.9.0.

Although it is unlikely that ClusterIngresses were correctly supported
given our use of the namespace attribute, the prior commit makes
referencing the namespace of a cluster-scoped object an error, and so
we're guaranteed to not be processing them correctly.

Additionally, we have only been looking at the "ingress.class"
annotation to determine whether we should process a
ClusterIngress/Ingress. This annotation first made an appearance in
Knative Serving 0.10.0 (after ClusterIngresses were removed), so we can
say that is our de facto minimum supported version anyway.