(from AES) Fix two regressions with internal vs Envoy cluster naming
- Use the envoy_name when generating the tracing configuration for
bootstrap, since we will later generate a cluster object with that
name after bootstrapping, as part of Ambex's config delivery via xDS.
- Pass the envoy_name to the diagnostics page rendering code for
Ambassador "Services" (aka tracing service, auth service, etc). We need
the envoy_name here because it is used as a key into a map that uses
envoy_name for estats lookups.

Adds a test for long cluster names for the TracingService, verifying
that we have correct, existing functionality when the service name is
long, and that we can load the diagnostics page correctly with HTTP 200
in this case.

refs https://github.com/datawire/ambassador/issues/3019
refs https://github.com/datawire/ambassador/issues/3021