(from AES) builder.mk: Fix an ordering issue in the build
@esmet showed me a real weird build error he got

    printf '%s\n' $(cat docker/base-envoy.docker) aes.local/base-envoy | /home/esmet/git/datawire/apro/ambassador/build-aux/bin/write-dockertagfile docker/base-envoy.docker.tag.local
    Error response from daemon: No such image: datawire/ambassador-base:envoy-1.6cfbd83d64807d25fe175c196ff15baf9ce4f3cb.opt
    make: *** [/home/esmet/git/datawire/apro/ambassador/build-aux/docker.mk:270: docker/base-envoy.docker.tag.local] Error 1

I went from "how is that happening?" to "oh, that's how it's happening...
how was it working for anyone else?" to finally understanding it.

When I wrote the initial version of the `docker/base-envoy.docker.stamp`
rule, I'd made the assumption that I could get away with not actually
pulling the image, because all of the `docker` commands would implicitly
do it lazily on-demand.  However, that's not true for the
`docker/base-envoy.docker.tag.*` rules (which, of everything that use
base-envoy.docker.stamp, is the first one that will trigger).  So, how was
it working at all?  Because `docker/container.txt.stamp` will implicitly
do the pull (even though it has no dependency relationship with
`base-envoy.docker` that Make knows about).  And for most runs, it just so
happened that Make decided to do `container.txt` before
`base-envoy.docker` so it all just worked out.

So, the actual change to make: Begone with that bad assumption that
`docker/base-envoy.docker` doesn't need to actually pull the image.