(from AES) test_knative: Fix
And also improve an error message that confused me when I was debugging
this.

This test breakage was introduced in commit 233264aca ("Move service and
endpoints processing to a separate processor", 2020-06-02) (merged in
PR #2761, 2020-10-20), but was not noticed because the knative tests are
skipped when using a too-old version of Kubernetes, and ambassador.git CI
is still using Kubernaut which has Kubernetes 1.10.