Add circuit_breakers to AuthService extauth cluster
In Ambassador we can add circuit breakers to Mappings directly,
and also in the Ambassador module in order to apply the settings to all mappings.
Since the AuthService is not a Mapping, that logic does not apply to the extauth
cluster. extauth is an Envoy cluster which also accepts custom circuit breaker settings.

In this PR I add support for setting circuit_breakers in the AuthService CRD.
The global circuit breakers configured in the Ambassador module are applied to extauth aswell.

Updated CRDs