Use a resource processor for ingresses
This change moves the handlers for the ingress classes and ingresses in
the fetcher to their own resource processors. Because of the strange
interdependence of the Ambassador service, which must be discovered,
ingress classes, and ingresses, we also introduce a dependency handling
mechanism that (a) makes it clear where that information is coming from
and (b) automatically generates the desired handling order for data
coming in from watt.