Add processor dependency management
Frequently, one or more processors will depend on something discovered
by another processor to work correctly. This had been implemented by
manually keeping track of these dependencies and putting them into a
list in the fetcher so that the order would be known in advance.
However, this approach is brittle and it's hard to find a good place to
store the dependencies that makes sense to every dependent processor.

This change adds support for dependency management, where a processor
may provide or want a specific set of data associated with the
processing of a given type (expressed as a Watt key). The ordering is
then automatically computed by traversing a graph of the dependencies.