Convert Knative Ingress paths from prefix match to path match
This change makes the paths specified in each Knative Ingress rule
prefix-matching. The previous behavior required the path pattern to
match the entire actual path. With a default of `/`, this in particular
meant that any path other than the root would not match, which
explicitly deviates from the "catch all" requirement in the Knative
spec.