fix: set diagd gunicorn keepalive to 300s
This is to help avoid 503's on the /ambassador/v0/check_alive
and similar endpoints. Currently, use-cases that probe this
health check are subject to race-conditions where gunicorn
closes idle TCP connections after 2s and does not send a FIN or RST to envoy.
Envoy, believing the tcp connection is still active, attempts to route
the check_alive request to gunicorn, but receives a RST as the TCP connection
is no longer valid according to gunicorn. This results in a failing
health check even though gunicorn is 'up'. This false-positive failure
mode is problematic especially in situations where one is leveraging
this health check to detect backend health from a CDN or other downstream load balancer.

Setting the keepalive timeout to 300s should be a safe default as typical
implementations will probably be polling this health endpoint more than once
every 5 minutes.