Upgrade envoy-control-plane v0.9.8→$(make guess-envoy-go-control-plane-commit)
A bunch of the methods that formerly returned an ecp_v2_cache.Resource
interface now return a boxed wrapper struct (ecp_v2_cache.ResourceWithTtl),
so we need to unbox that by adding ".Resource" in a bunch of places.

Also, it introduced some code that the linter is unhappy with, so add and
another exception for it.

Signed-off-by: Luke Shumaker <lukeshu@datawire.io>
(cherry picked from commit 27b823ef5179f2249f6953ccd5a6d4ddd954b820)