CRDs: v2: roundtrip_test.go: Switch to YAML, part 1
YAML is easier to work with than JSON.  This starts by leaving the JSON
in place, and comparing the YAML to the JSON to ensure that I didn't break
anything when converting it from JSON to YAML (using `sed` and some Emacs
macros).

This closely mimics 05f9dff09 (v3alpha1 CRDs: roundtrip_test.go:
Switch to YAML, part 1, 2021-10-05).

Signed-off-by: Luke Shumaker <lukeshu@datawire.io>