Prep for v2 CRDs: Handle non-structural CRD schemas as best we can [ci-skip]
In order to support 2.1's better migration story, the coming commits
will be adding back in the getambassador.io/v2 CRDs.  Unfortunately,
these CRDs have schemas that are non-structural, which will by
rejected by both the apiserver and pkg/kates.NewValidator()'s call to
k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/validation.ValidateCustomResourceDefinition().

For the apiserver, the solution is clear but terrible: Munge the
schema to the degenerate case, such that the apiserver will reject it.

For our own internal validator, that is not an acceptable solution.
Previously we had been feeding the same `ambassador-crds.yaml` file to
both the apiserver and to our internal validator; so in order to have
a different solution, we need to have a different YAML file: enter
`cmd/entrypoint/crds.yaml`.  And in this version, the anwer is: Don't
munge the schemas, and get rid of the call to
ValidateCustomResourceDefinition() in order to make that acceptable.

This is marked [ci-skip] because it needs a `make go-mod-tidy && make
generate`.

Signed-off-by: Aidan Hahn <aidanhahn@datawire.io>
Signed-off-by: Luke Shumaker <lukeshu@datawire.io>