Prep for v2 CRDs: cmd/entrypoint: List all server resource versions, not just preferred versions
This should have been done in 572f9848f (Get real persnickety about the
resource versions that we watch, 2021-09-27).  Before that commit, it
didn't care about multiple versions, so it was OK to only get the
apiserver's preferred versions.  But now it does care about different
versions, so it should be listing all of them.

In the 2.0 release chain, this turns out to not matter much, because the
only version of the getambassador.io resources is v3alpha1.  But in 2.1
there will be a better migration story, so the older versions will exist
as well.  And v2 will actually be the apiserver's preferred version, and
so if we only list the apiserver's preferred versions, then Emissary will
bork because it thinks the v3alpha1 versions don't exist at all!

Signed-off-by: Luke Shumaker <lukeshu@datawire.io>