builder.mk, release.mk: s/\$\((BUILD|RELEASE)_VERSION\)/$(patsubst v%,%,$(VERSION))/g
This isn't the end game.  If you're just looking at this commit, you
may think "why not define `VERSION_NO_V = $(patsubst v%,%,$(VERSION))`
and use that everywhere, instead of repeating the patsubst a million
times?"  Well, subsequent commits will be getting rid of most of them.
This is just the semi-automated step to have the next commit be more
readable.  And then the patsubst that remain should be painful and
verbose, because they are indicators of remaining inconsistency in the
build system.

Signed-off-by: Luke Shumaker <lukeshu@datawire.io>