Validate annotations in the snapshot; merge `annotation.go`s
I must apologize that this is a bit of a monster of a commit.

This is in prep for the annotation-unfolding to happen purely in Go.

 1. Smash cmd/entrypoint/annotations.go in to
    pkg/snapshot/v1/annotations.go (and move annotaations_test.go)
 2. Have it validate the annotations when doing so.  This means packing
    more metadata in to the snapshot.
 3. Have the snapshot know where annotations came from (so that the Python
    will eventually be able to use that info to generate the rkey)

In annotations_test.go I had to add a few `service: quote:80` lines to
Mappings, in order to get the validation to pass.  Similarly, I had to
upgrade a few resources in consul_test.go from getambassador.io/v1 to
getambassador.io/v2; as the validator correctly complained:

    unsupported GroupVersionKind "getambassador.io/v1, Kind=KubernetesServiceResolver", ignoring

Signed-off-by: Luke Shumaker <lukeshu@datawire.io>