CRDs: Fix being able to have empty-but-set lists and dicts [ci-skip]
One of the reasons that
https://github.com/emissary-ingress/emissary/pull/4016 is failing is
that the Go CRD definitions are failing to preserve set-but-empty
`add_request_headers: {}` and `remove_request_headers: []` fields.
Preserving these is important because if the `ambassador` `Module`
sets a non-empty global default, it is desirable to be able to turn
that off locally for a Mapping.

So, since both a nil slice and an empty slice trigger `,omitempty`, I
just added `*` in front of the types to change what "empty" means.
Unfortunately, this causes both `conversion-gen` and `controller-gen`
to generate invalid code.

So I went to work patching both `conversion-gen` and `controller-gen`.
As much as I hate checking in `vendor/`, I have to admit that in this
instance it makes the PR easier to review, since I don't have to link
to a diff in an external repo.  So, yeah, unlike most PRs you actually
should pay attention to the changes in `vendor/` this time.

Previously, sometimes when `conversion-gen` encountered 2 types it
couldn't convert, `conversion-gen` would emit

    compileErrorOnMissingConversion()

(triggering a compile error), and sometimes it would emit

    // WARNING: in.StructMemberName requires manual conversion: inconvertible types (Type1 vs Type2)

and I'd have told you that I had absolutely zero intuition for when it
would do one vs the other.  Well, with this latest change it always
does `compileErrorOnMissingConversion()`, so I had to add a few

    // +k8s:conversion-gen=false

comments to prevent compile errors.

This commit only includes my human-authored changes; for review
purposes I'm holding the `make generate`-authored changes for a
separate commit.  I'm not sure if in this situation that's actually
helpful to the reviewer, but often it is; so I'm playing it safe.

Signed-off-by: Luke Shumaker <lukeshu@datawire.io>